- bug fix in return type of byte_swap32() - -Wpedantic is a bit too much for now

+++ src/msgpack.h
@@ -365,24 +365,24 @@ namespace msgpackcpp
         return static_cast<uint16_t>(((v & 0x00FF) << 8) | ((v & 0xFF00) >> 8));
     }
 
-    constexpr uint64_t byte_swap32(uint32_t v)
+    constexpr uint32_t byte_swap32(uint32_t v)
     {
         return static_cast<uint32_t>(((v & 0x000000FFu) << 24) |
-                                        ((v & 0x0000FF00u) << 8)  |
-                                        ((v & 0x00FF0000u) >> 8)  |
-                                        ((v & 0xFF000000u) >> 24));
+                                     ((v & 0x0000FF00u) << 8)  |
+                                     ((v & 0x00FF0000u) >> 8)  |
+                                     ((v & 0xFF000000u) >> 24));
     }
 
     constexpr uint64_t byte_swap64(uint64_t v)
     {
         return static_cast<uint64_t>(((v & 0x00000000000000FFULL) << 56) |
-                                    ((v & 0x000000000000FF00ULL) << 40) |
-                                    ((v & 0x0000000000FF0000ULL) << 24) |
-                                    ((v & 0x00000000FF000000ULL) << 8)  |
-                                    ((v & 0x000000FF00000000ULL) >> 8)  |
-                                    ((v & 0x0000FF0000000000ULL) >> 24) |
-                                    ((v & 0x00FF000000000000ULL) >> 40) |
-                                    ((v & 0xFF00000000000000ULL) >> 56));
+                                     ((v & 0x000000000000FF00ULL) << 40) |
+                                     ((v & 0x0000000000FF0000ULL) << 24) |
+                                     ((v & 0x00000000FF000000ULL) << 8)  |
+                                     ((v & 0x000000FF00000000ULL) >> 8)  |
+                                     ((v & 0x0000FF0000000000ULL) >> 24) |
+                                     ((v & 0x00FF000000000000ULL) >> 40) |
+                                     ((v & 0xFF00000000000000ULL) >> 56));
     }
 
     static_assert(byte_swap16(0x1234)               == 0x3412,              "bad swap");
